# examples/demo-repo/src/main/resources/openapi/customer-onboarding.openapi.yaml
openapi: 3.0.3
info:
  title: Customer Onboarding API
  version: 1.0.0
paths:
  /customers/onboard:
    post:
      operationId: startCustomerOnboarding
      summary: Start onboarding for a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardingRequest'
            examples:
              validCustomer:
                summary: Customer with referral code
                value:
                  email: jane.doe@example.com
                  firstName: Jane
                  lastName: Doe
                  birthDate: '1990-04-12'
                  referralCode: PARTNER-42
      responses:
        '202':
          description: Onboarding accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingResponse'
              examples:
                accepted:
                  value:
                    customerId: c-123
                    status: ACTIVE
                    riskScore: 20
                    kycReference: kyc-789
        '409':
          description: Customer already exists
  /customers/{customerId}/status:
    get:
      operationId: getCustomerOnboardingStatus
      summary: Return onboarding status for a customer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current customer status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardingStatusResponse'
              examples:
                manualReview:
                  value:
                    customerId: c-123
                    status: MANUAL_REVIEW
                    reviewReason: KYC rejected or risk score above threshold
components:
  schemas:
    OnboardingRequest:
      type: object
      required: [email, firstName, lastName, birthDate]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        referralCode:
          type: string
    OnboardingResponse:
      type: object
      properties:
        customerId:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, MANUAL_REVIEW]
        riskScore:
          type: integer
        kycReference:
          type: string
    OnboardingStatusResponse:
      type: object
      properties:
        customerId:
          type: string
        status:
          type: string
        reviewReason:
          type: string